DinnerDash/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cloudinary==1.29.0
coverage==6.4.2
Django==4.0.6
django-cors-headers==3.13.0
django-multiselectfield==0.1.12
django-multiupload==0.6.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
environ==1.0
factory-boy==3.2.1
Faker==13.15.1
Gears==0.7.2
glob2==0.7
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
whitenoise==6.2.0
